Remove GPU implementation of Stokesian Dynamics
