+++ src/core/CMakeLists.txt
@@ -108,12 +108,7 @@ if(WITH_UNIT_TESTS)
   add_subdirectory(unit_tests)
 endif(WITH_UNIT_TESTS)
 
-if(STOKESIAN_DYNAMICS OR STOKESIAN_DYNAMICS_GPU)
+if(STOKESIAN_DYNAMICS)
   add_subdirectory(stokesian_dynamics)
-  if(STOKESIAN_DYNAMICS)
-    target_link_libraries(EspressoCore PRIVATE StokesianDynamics::sd_cpu)
-  endif()
-  if(STOKESIAN_DYNAMICS_GPU)
-    target_link_libraries(EspressoCore PRIVATE StokesianDynamics::sd_gpu)
-  endif()
+  target_link_libraries(EspressoCore PRIVATE StokesianDynamics::sd_cpu)
 endif()
