+++ src/core/stokesian_dynamics/sd_interface.hpp
@@ -32,14 +32,10 @@
 
 #include <boost/mpi/communicator.hpp>
 
-#include <string>
 #include <unordered_map>
 
 void set_sd_viscosity(double eta);
 double get_sd_viscosity();
-
-void set_sd_device(std::string const &dev);
-std::string get_sd_device();
 
 void set_sd_radius_dict(std::unordered_map<int, double> const &x);
 std::unordered_map<int, double> get_sd_radius_dict();
