+++ src/python/espressomd/integrate.pxd
@@ -43,12 +43,9 @@ IF NPT:
                                              cbool zdir_rescale, cbool cubic_box)
 
 cdef extern from "stokesian_dynamics/sd_interface.hpp":
-    IF(STOKESIAN_DYNAMICS or STOKESIAN_DYNAMICS_GPU):
+    IF STOKESIAN_DYNAMICS:
         void set_sd_viscosity(double eta)
         double get_sd_viscosity()
-
-        void set_sd_device(const string & dev)
-        string get_sd_device()
 
         void set_sd_radius_dict(const unordered_map[int, double] & radius_dict)
         unordered_map[int, double] get_sd_radius_dict()
@@ -56,7 +53,7 @@ cdef extern from "stokesian_dynamics/sd_interface.hpp":
         void set_sd_flags(int flg)
         int get_sd_flags()
 
-IF(STOKESIAN_DYNAMICS or STOKESIAN_DYNAMICS_GPU):
+IF STOKESIAN_DYNAMICS:
     cpdef enum flags:
         NONE = 0,
         SELF_MOBILITY = 1 << 0,
