+++ testsuite/scripts/samples/CMakeLists.txt
@@ -26,8 +26,6 @@ sample_test(FILE test_chamber_game.py)
 sample_test(FILE test_constraints.py)
 sample_test(FILE test_dancing.py SUFFIX cpu_ft)
 sample_test(FILE test_dancing.py SUFFIX cpu_fts)
-sample_test(FILE test_dancing.py SUFFIX gpu_ft LABELS "gpu")
-sample_test(FILE test_dancing.py SUFFIX gpu_fts LABELS "gpu")
 sample_test(FILE test_diffusion_coefficient.py)
 sample_test(FILE test_dpd.py)
 sample_test(FILE test_drude_bmimpf6.py SUFFIX cpu)
